+++ src/game/actors/attributes/LuckAttribute.java
@@ -2,7 +2,7 @@ package actors.attributes;
 
 public class LuckAttribute extends Attribute {
 
-    public LuckAttribute(int value) {
+    public LuckAttribute(double value) {
         super(AttributeTypes.LUCK, value);
     }
 }
