+++ src/game/actors/resistances/BludgeoningResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class BludgeoningResistance extends Resistance {
-    public BludgeoningResistance(int value) {
+    public BludgeoningResistance(double value) {
         super(DamageTypes.BLUDGEONING, value);
     }
 }
