+++ src/game/actors/resistances/EarthResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class EarthResistance extends Resistance {
-    public EarthResistance(int value) {
+    public EarthResistance(double value) {
         super(DamageTypes.EARTH, value);
     }
 }
