+++ src/game/actors/resistances/LightningResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class LightningResistance extends Resistance {
-    public LightningResistance(int value) {
+    public LightningResistance(double value) {
         super(DamageTypes.LIGHTNING, value);
     }
 }
