+++ src/game/actors/resistances/PiercingResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class PiercingResistance extends Resistance {
-    public PiercingResistance(int value) {
+    public PiercingResistance(double value) {
         super(DamageTypes.PIERCING, value);
     }
 }
