+++ src/game/actors/resistances/VenomResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class VenomResistance extends Resistance {
-    public VenomResistance(int value) {
+    public VenomResistance(double value) {
         super(DamageTypes.VENOM, value);
     }
 }
