+++ src/game/actors/resources/ManaValues.java
@@ -1,7 +1,7 @@
 package actors.resources;
 
 public class ManaValues extends ResourceValues {
-    public ManaValues(int maxValue, int regenValue) {
+    public ManaValues(double maxValue, double regenValue) {
         super(maxValue, maxValue, regenValue);
     }
